+++ channel/apiv1/cloud_channel_client.go
@@ -72,6 +72,9 @@ type CloudChannelCallOptions struct {
 	ListOffers                      []gax.CallOption
 	ListPurchasableSkus             []gax.CallOption
 	ListPurchasableOffers           []gax.CallOption
+	RegisterSubscriber              []gax.CallOption
+	UnregisterSubscriber            []gax.CallOption
+	ListSubscribers                 []gax.CallOption
 }
 
 func defaultCloudChannelClientOptions() []option.ClientOption {
@@ -308,6 +311,39 @@ func defaultCloudChannelCallOptions() *CloudChannelCallOptions {
 				})
 			}),
 		},
+		RegisterSubscriber: []gax.CallOption{
+			gax.WithRetry(func() gax.Retryer {
+				return gax.OnCodes([]codes.Code{
+					codes.Unavailable,
+				}, gax.Backoff{
+					Initial:    1000 * time.Millisecond,
+					Max:        10000 * time.Millisecond,
+					Multiplier: 1.30,
+				})
+			}),
+		},
+		UnregisterSubscriber: []gax.CallOption{
+			gax.WithRetry(func() gax.Retryer {
+				return gax.OnCodes([]codes.Code{
+					codes.Unavailable,
+				}, gax.Backoff{
+					Initial:    1000 * time.Millisecond,
+					Max:        10000 * time.Millisecond,
+					Multiplier: 1.30,
+				})
+			}),
+		},
+		ListSubscribers: []gax.CallOption{
+			gax.WithRetry(func() gax.Retryer {
+				return gax.OnCodes([]codes.Code{
+					codes.Unavailable,
+				}, gax.Backoff{
+					Initial:    1000 * time.Millisecond,
+					Max:        10000 * time.Millisecond,
+					Multiplier: 1.30,
+				})
+			}),
+		},
 	}
 }
 
@@ -438,7 +474,7 @@ func (c *CloudChannelClient) setGoogleClientInfo(keyval ...string) {
 //   request.
 //
 // Return Value:
-//  List of Customers pertaining to the reseller or empty list if
+// List of Customers pertaining to the reseller or empty list if
 // there are none.
 func (c *CloudChannelClient) ListCustomers(ctx context.Context, req *channelpb.ListCustomersRequest, opts ...gax.CallOption) *CustomerIterator {
 	md := metadata.Pairs("x-goog-request-params", fmt.Sprintf("%s=%v", "parent", url.QueryEscape(req.GetParent())))
@@ -494,7 +530,7 @@ func (c *CloudChannelClient) ListCustomers(ctx context.Context, req *channelpb.L
 //   the result of an invalid name parameter.
 //
 // Return Value:
-//  Customer resource if found, error otherwise.
+// Customer resource if found, error otherwise.
 func (c *CloudChannelClient) GetCustomer(ctx context.Context, req *channelpb.GetCustomerRequest, opts ...gax.CallOption) (*channelpb.Customer, error) {
 	if _, ok := ctx.Deadline(); !ok && !c.disableDeadlines {
 		cctx, cancel := context.WithTimeout(ctx, 60000*time.Millisecond)
@@ -529,12 +565,12 @@ func (c *CloudChannelClient) GetCustomer(ctx context.Context, req *channelpb.Get
 //
 //   INVALID_VALUE: Invalid domain value in the request.
 //
-//   NOT_FOUND: If there is no CloudIdentityCustomerAccount customer
-//   for the domain specified in the request.
-//
 // Return Value:
-//  List of CloudIdentityCustomerAccount resources if any exist for
-// the domain, otherwise an error is returned.
+// List of CloudIdentityCustomerAccount resources for the domain.
+// List may be empty.
+//
+// Note: in the v1alpha1 version of the API, a NOT_FOUND error is returned if
+// no CloudIdentityCustomerAccount resources match the domain.
 func (c *CloudChannelClient) CheckCloudIdentityAccountsExist(ctx context.Context, req *channelpb.CheckCloudIdentityAccountsExistRequest, opts ...gax.CallOption) (*channelpb.CheckCloudIdentityAccountsExistResponse, error) {
 	if _, ok := ctx.Deadline(); !ok && !c.disableDeadlines {
 		cctx, cancel := context.WithTimeout(ctx, 60000*time.Millisecond)
@@ -561,8 +597,18 @@ func (c *CloudChannelClient) CheckCloudIdentityAccountsExist(ctx context.Context
 //
 // Possible Error Codes:
 //
+//   PERMISSION_DENIED: If the reseller account making the request and the
+//   reseller account being queried for are different.
+//
+//   INVALID_ARGUMENT: It can happen in following scenarios -
+//
+//     Missing or invalid required parameters in the request.
+//
+//     Domain field value doesn’t match the domain specified in primary
+//     email.
+//
 // Return Value:
-//  If successful, the newly created Customer resource, otherwise
+// If successful, the newly created Customer resource, otherwise
 // returns an error.
 func (c *CloudChannelClient) CreateCustomer(ctx context.Context, req *channelpb.CreateCustomerRequest, opts ...gax.CallOption) (*channelpb.Customer, error) {
 	if _, ok := ctx.Deadline(); !ok && !c.disableDeadlines {
@@ -600,7 +646,7 @@ func (c *CloudChannelClient) CreateCustomer(ctx context.Context, req *channelpb.
 //   specified in the request.
 //
 // Return Value:
-//  If successful, the updated Customer resource, otherwise returns
+// If successful, the updated Customer resource, otherwise returns
 // an error.
 func (c *CloudChannelClient) UpdateCustomer(ctx context.Context, req *channelpb.UpdateCustomerRequest, opts ...gax.CallOption) (*channelpb.Customer, error) {
 	if _, ok := ctx.Deadline(); !ok && !c.disableDeadlines {
@@ -675,7 +721,7 @@ func (c *CloudChannelClient) DeleteCustomer(ctx context.Context, req *channelpb.
 //   Contact Cloud Channel support in this case.
 //
 // Return Value:
-//   Long Running Operation ID.
+// Long Running Operation ID.
 //
 // To get the results of the operation, call the GetOperation method of
 // CloudChannelOperationsService. The Operation metadata will contain an
@@ -712,7 +758,7 @@ func (c *CloudChannelClient) ProvisionCloudIdentity(ctx context.Context, req *ch
 //   INVALID_ARGUMENT: Missing or invalid required parameters in the request.
 //
 // Return Value:
-//  List of Entitlements belonging to the customer, or empty list if
+// List of Entitlements belonging to the customer, or empty list if
 // there are none.
 func (c *CloudChannelClient) ListEntitlements(ctx context.Context, req *channelpb.ListEntitlementsRequest, opts ...gax.CallOption) *EntitlementIterator {
 	md := metadata.Pairs("x-goog-request-params", fmt.Sprintf("%s=%v", "parent", url.QueryEscape(req.GetParent())))
@@ -763,8 +809,19 @@ func (c *CloudChannelClient) ListEntitlements(ctx context.Context, req *channelp
 //
 // Possible Error Codes:
 //
+//   PERMISSION_DENIED: Appears because of one of the following -
+//
+//     The customer doesn’t belong to the reseller and no auth token.
+//
+//     The supplied auth token is invalid.
+//
+//     The reseller account making the request and the queries reseller
+//     account are different.
+//
+//   INVALID_ARGUMENT: Missing or invalid required parameters in the request.
+//
 // Return Value:
-//  List of TransferableSku for the given customer.
+// List of TransferableSku for the given customer.
 func (c *CloudChannelClient) ListTransferableSkus(ctx context.Context, req *channelpb.ListTransferableSkusRequest, opts ...gax.CallOption) *TransferableSkuIterator {
 	md := metadata.Pairs("x-goog-request-params", fmt.Sprintf("%s=%v", "parent", url.QueryEscape(req.GetParent())))
 	ctx = insertMetadata(ctx, c.xGoogMetadata, md)
@@ -814,10 +871,13 @@ func (c *CloudChannelClient) ListTransferableSkus(ctx context.Context, req *chan
 //
 // Possible Error Codes:
 //
-//   PERMISSION_DENIED, due to one of the following reasons: (a) If the
-//   customer doesn’t belong to the reseller and no auth token or invalid auth
-//   token is supplied. (b) If the reseller account making the request and the
-//   reseller account being queried for are different.
+//   PERMISSION_DENIED: Appears because of one of the following:
+//
+//     If the customer doesn’t belong to the reseller and no auth token or
+//     invalid auth token is supplied.
+//
+//     If the reseller account making the request and the reseller account
+//     being queried for are different.
 //
 //   INVALID_ARGUMENT: Missing or invalid required parameters in the
 //   request.
@@ -876,7 +936,7 @@ func (c *CloudChannelClient) ListTransferableOffers(ctx context.Context, req *ch
 //   NOT_FOUND: If the entitlement is not found for the customer.
 //
 // Return Value:
-//  If found, the requested Entitlement resource, otherwise returns
+// If found, the requested Entitlement resource, otherwise returns
 // an error.
 func (c *CloudChannelClient) GetEntitlement(ctx context.Context, req *channelpb.GetEntitlementRequest, opts ...gax.CallOption) (*channelpb.Entitlement, error) {
 	if _, ok := ctx.Deadline(); !ok && !c.disableDeadlines {
@@ -903,8 +963,53 @@ func (c *CloudChannelClient) GetEntitlement(ctx context.Context, req *channelpb.
 //
 // Possible Error Codes:
 //
+//   PERMISSION_DENIED: If the customer doesn’t belong to the reseller.
+//
+//   INVALID_ARGUMENT: It can happen in below scenarios -
+//
+//     Missing or invalid required parameters in the request.
+//
+//     Cannot purchase an entitlement if there is already an entitlement for
+//     customer, for a SKU from the same product family.
+//
+//     INVALID_VALUE: Offer passed in isn’t valid. Make sure OfferId is
+//     valid. If it is valid, then contact Google Channel support for further
+//     troubleshooting.
+//
+//   NOT_FOUND: If the customer or offer resource is not found for the
+//   reseller.
+//
+//   ALREADY_EXISTS: This failure can happen in the following cases:
+//
+//     If the SKU has been already purchased for the customer.
+//
+//     If the customer’s primary email already exists. In this case retry
+//     after changing the customer’s primary contact email.
+//
+//   CONDITION_NOT_MET or FAILED_PRECONDITION: This failure can happen in the
+//   following cases:
+//
+//     Purchasing a SKU that requires domain verification and the domain has
+//     not been verified.
+//
+//     Purchasing an Add-On SKU like Vault or Drive without purchasing the
+//     pre-requisite SKU, such as Google Workspace Business Starter.
+//
+//     Applicable only for developer accounts: reseller and resold domain.
+//     Must meet the following domain naming requirements:
+//
+//       Domain names must start with goog-test.
+//
+//       Resold domain names must include the reseller domain.
+//
+//   INTERNAL: Any non-user error related to a technical issue in the
+//   backend. Contact Cloud Channel Support in this case.
+//
+//   UNKNOWN: Any non-user error related to a technical issue in the
+//   backend. Contact Cloud Channel Support in this case.
+//
 // Return Value:
-//  Long Running Operation ID.
+// Long Running Operation ID.
 //
 // To get the results of the operation, call the GetOperation method of
 // CloudChannelOperationsService. The Operation metadata will contain an
@@ -955,7 +1060,7 @@ func (c *CloudChannelClient) CreateEntitlement(ctx context.Context, req *channel
 //   In this case, contact Cloud Channel support.
 //
 // Return Value:
-//  Long Running Operation ID.
+// Long Running Operation ID.
 //
 // To get the results of the operation, call the GetOperation method of
 // CloudChannelOperationsService. The Operation metadata will contain an
@@ -1007,7 +1112,7 @@ func (c *CloudChannelClient) ChangeParameters(ctx context.Context, req *channelp
 //   In this case, contact Cloud Channel support.
 //
 // Return Value:
-//  Long Running Operation ID.
+// Long Running Operation ID.
 //
 // To get the results of the operation, call the GetOperation method of
 // CloudChannelOperationsService. The Operation metadata will contain an
@@ -1056,7 +1161,7 @@ func (c *CloudChannelClient) ChangeRenewalSettings(ctx context.Context, req *cha
 //   In this case, contact Cloud Channel support.
 //
 // Return Value:
-//  Long Running Operation ID.
+// Long Running Operation ID.
 //
 // To get the results of the operation, call the GetOperation method of
 // CloudChannelOperationsService. The Operation metadata will contain an
@@ -1109,7 +1214,7 @@ func (c *CloudChannelClient) ChangeOffer(ctx context.Context, req *channelpb.Cha
 //   in the backend. In this case, contact Cloud Channel support.
 //
 // Return Value:
-//  Long Running Operation ID.
+// Long Running Operation ID.
 //
 // To get the results of the operation, call the GetOperation method of
 // CloudChannelOperationsService. The Operation metadata will contain an
@@ -1158,7 +1263,7 @@ func (c *CloudChannelClient) StartPaidService(ctx context.Context, req *channelp
 //   In this case, contact Cloud Channel support.
 //
 // Return Value:
-//  Long Running Operation ID.
+// Long Running Operation ID.
 //
 // To get the results of the operation, call the GetOperation method of
 // CloudChannelOperationsService. The Operation metadata will contain an
@@ -1213,7 +1318,7 @@ func (c *CloudChannelClient) SuspendEntitlement(ctx context.Context, req *channe
 //   In this case, contact Cloud Channel support.
 //
 // Return Value:
-//  Long Running Operation ID.
+// Long Running Operation ID.
 //
 // To get the results of the operation, call the GetOperation method of
 // CloudChannelOperationsService. The response will contain
@@ -1273,7 +1378,7 @@ func (c *CloudChannelClient) CancelEntitlement(ctx context.Context, req *channel
 //   In this case, contact Cloud Channel support.
 //
 // Return Value:
-//  Long Running Operation ID.
+// Long Running Operation ID.
 //
 // To get the results of the operation, call the GetOperation method of
 // CloudChannelOperationsService. The Operation metadata will contain an
@@ -1305,8 +1410,41 @@ func (c *CloudChannelClient) ActivateEntitlement(ctx context.Context, req *chann
 //
 // Possible Error Codes:
 //
+//   PERMISSION_DENIED: If the customer doesn’t belong to the reseller.
+//
+//   INVALID_ARGUMENT: Missing or invalid required parameters in the request.
+//
+//   NOT_FOUND: If the customer or offer resource is not found for the
+//   reseller.
+//
+//   ALREADY_EXISTS: If the SKU has been already transferred for the customer.
+//
+//   CONDITION_NOT_MET or FAILED_PRECONDITION: This failure can happen in the
+//   following cases:
+//
+//     Transferring a SKU that requires domain verification and the domain
+//     has not been verified.
+//
+//     Transferring an Add-On SKU like Vault or Drive without transferring
+//     the pre-requisite SKU, such as G Suite Basic.
+//
+//     Applicable only for developer accounts: reseller and resold domain
+//     must follow the domain naming convention as follows:
+//
+//       Domain names must start with goog-test.
+//
+//       Resold domain names must include the reseller domain.
+//
+//     All transferring entitlements must be specified.
+//
+//   INTERNAL: Any non-user error related to a technical issue in the backend.
+//   Please contact Cloud Channel Support in this case.
+//
+//   UNKNOWN: Any non-user error related to a technical issue in the backend.
+//   Please contact Cloud Channel Support in this case.
+//
 // Return Value:
-//  Long Running Operation ID.
+// Long Running Operation ID.
 //
 // To get the results of the operation, call the GetOperation method of
 // CloudChannelOperationsService. The Operation metadata will contain an
@@ -1338,8 +1476,39 @@ func (c *CloudChannelClient) TransferEntitlements(ctx context.Context, req *chan
 //
 // Possible Error Codes:
 //
+//   PERMISSION_DENIED: If the customer doesn’t belong to the reseller.
+//
+//   INVALID_ARGUMENT: Missing or invalid required parameters in the request.
+//
+//   NOT_FOUND: If the customer or offer resource is not found for the
+//   reseller.
+//
+//   ALREADY_EXISTS: If the SKU has been already transferred for the customer.
+//
+//   CONDITION_NOT_MET or FAILED_PRECONDITION: This failure can happen in
+//   the following cases:
+//
+//     Transferring a SKU that requires domain verification and the domain
+//     has not been verified.
+//
+//     Transferring an Add-On SKU like Vault or Drive without purchasing the
+//     pre-requisite SKU, such as G Suite Basic.
+//
+//     Applicable only for developer accounts: reseller and resold domain
+//     must follow the domain naming convention as follows:
+//
+//       Domain names must start with goog-test.
+//
+//       Resold domain names must include the reseller domain.
+//
+//   INTERNAL: Any non-user error related to a technical issue in the backend.
+//   Please contact Cloud Channel Support in this case.
+//
+//   UNKNOWN: Any non-user error related to a technical issue in the backend.
+//   Please contact Cloud Channel Support in this case.
+//
 // Return Value:
-//  Long Running Operation ID.
+// Long Running Operation ID.
 //
 // To get the results of the operation, call the GetOperation method of
 // CloudChannelOperationsService. The response will contain
@@ -1380,7 +1549,7 @@ func (c *CloudChannelClient) TransferEntitlementsToGoogle(ctx context.Context, r
 //   request.
 //
 // Return Value:
-//  If successful, returns the list of ChannelPartnerLink resources
+// If successful, returns the list of ChannelPartnerLink resources
 // for the distributor account, otherwise returns an error.
 func (c *CloudChannelClient) ListChannelPartnerLinks(ctx context.Context, req *channelpb.ListChannelPartnerLinksRequest, opts ...gax.CallOption) *ChannelPartnerLinkIterator {
 	md := metadata.Pairs("x-goog-request-params", fmt.Sprintf("%s=%v", "parent", url.QueryEscape(req.GetParent())))
@@ -1437,7 +1606,7 @@ func (c *CloudChannelClient) ListChannelPartnerLinks(ctx context.Context, req *c
 //   due invalid channel partner link name.
 //
 // Return Value:
-//  ChannelPartnerLink resource if found, otherwise returns an error.
+// ChannelPartnerLink resource if found, otherwise returns an error.
 func (c *CloudChannelClient) GetChannelPartnerLink(ctx context.Context, req *channelpb.GetChannelPartnerLinkRequest, opts ...gax.CallOption) (*channelpb.ChannelPartnerLink, error) {
 	if _, ok := ctx.Deadline(); !ok && !c.disableDeadlines {
 		cctx, cancel := context.WithTimeout(ctx, 60000*time.Millisecond)
@@ -1486,7 +1655,7 @@ func (c *CloudChannelClient) GetChannelPartnerLink(ctx context.Context, req *cha
 //   the backend. In this case, contact Cloud Channel support.
 //
 // Return Value:
-//  Newly created ChannelPartnerLink resource if successful,
+// Newly created ChannelPartnerLink resource if successful,
 // otherwise error is returned.
 func (c *CloudChannelClient) CreateChannelPartnerLink(ctx context.Context, req *channelpb.CreateChannelPartnerLinkRequest, opts ...gax.CallOption) (*channelpb.ChannelPartnerLink, error) {
 	if _, ok := ctx.Deadline(); !ok && !c.disableDeadlines {
@@ -1515,8 +1684,28 @@ func (c *CloudChannelClient) CreateChannelPartnerLink(ctx context.Context, req *
 //
 // Possible Error Codes:
 //
+//   PERMISSION_DENIED: If the reseller account making the request and the
+//   reseller account being queried for are different.
+//
+//   INVALID_ARGUMENT: It can happen in following scenarios -
+//
+//     Missing or invalid required parameters in the request.
+//
+//     Updating link state from invited to active or suspended.
+//
+//     Sending reseller_cloud_identity_id, invite_url or name in update
+//     mask.
+//
+//   NOT_FOUND: ChannelPartnerLink resource not found.
+//
+//   INTERNAL: Any non-user error related to a technical issue in the backend.
+//   In this case, contact Cloud Channel support.
+//
+//   UNKNOWN: Any non-user error related to a technical issue in the backend.
+//   In this case, contact Cloud Channel support.
+//
 // Return Value:
-//  If successful, the updated ChannelPartnerLink resource, otherwise
+// If successful, the updated ChannelPartnerLink resource, otherwise
 // returns an error.
 func (c *CloudChannelClient) UpdateChannelPartnerLink(ctx context.Context, req *channelpb.UpdateChannelPartnerLinkRequest, opts ...gax.CallOption) (*channelpb.ChannelPartnerLink, error) {
 	if _, ok := ctx.Deadline(); !ok && !c.disableDeadlines {
@@ -1765,6 +1954,159 @@ func (c *CloudChannelClient) ListPurchasableOffers(ctx context.Context, req *cha
 
 		it.Response = resp
 		return resp.GetPurchasableOffers(), resp.GetNextPageToken(), nil
+	}
+	fetch := func(pageSize int, pageToken string) (string, error) {
+		items, nextPageToken, err := it.InternalFetch(pageSize, pageToken)
+		if err != nil {
+			return "", err
+		}
+		it.items = append(it.items, items...)
+		return nextPageToken, nil
+	}
+	it.pageInfo, it.nextFunc = iterator.NewPageInfo(fetch, it.bufLen, it.takeBuf)
+	it.pageInfo.MaxSize = int(req.GetPageSize())
+	it.pageInfo.Token = req.GetPageToken()
+	return it
+}
+
+// RegisterSubscriber registers a service account with subscriber privileges on the Cloud Pub/Sub
+// topic created for this Channel Services account. Once you create a
+// subscriber, you will get the events as per SubscriberEvent
+//
+// Possible Error Codes:
+//
+//   PERMISSION_DENIED: If the reseller account making the request and the
+//   reseller account being provided are different, or if the impersonated user
+//   is not a super admin.
+//
+//   INVALID_ARGUMENT: Missing or invalid required parameters in the
+//   request.
+//
+//   INTERNAL: Any non-user error related to a technical issue in the
+//   backend. In this case, contact Cloud Channel support.
+//
+//   UNKNOWN: Any non-user error related to a technical issue in
+//   the backend. In this case, contact Cloud Channel support.
+//
+// Return Value:
+// Topic name with service email address registered if successful,
+// otherwise error is returned.
+func (c *CloudChannelClient) RegisterSubscriber(ctx context.Context, req *channelpb.RegisterSubscriberRequest, opts ...gax.CallOption) (*channelpb.RegisterSubscriberResponse, error) {
+	if _, ok := ctx.Deadline(); !ok && !c.disableDeadlines {
+		cctx, cancel := context.WithTimeout(ctx, 60000*time.Millisecond)
+		defer cancel()
+		ctx = cctx
+	}
+	md := metadata.Pairs("x-goog-request-params", fmt.Sprintf("%s=%v", "account", url.QueryEscape(req.GetAccount())))
+	ctx = insertMetadata(ctx, c.xGoogMetadata, md)
+	opts = append(c.CallOptions.RegisterSubscriber[0:len(c.CallOptions.RegisterSubscriber):len(c.CallOptions.RegisterSubscriber)], opts...)
+	var resp *channelpb.RegisterSubscriberResponse
+	err := gax.Invoke(ctx, func(ctx context.Context, settings gax.CallSettings) error {
+		var err error
+		resp, err = c.cloudChannelClient.RegisterSubscriber(ctx, req, settings.GRPC...)
+		return err
+	}, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return resp, nil
+}
+
+// UnregisterSubscriber unregisters a service account with subscriber privileges on the Cloud
+// Pub/Sub topic created for this Channel Services account. If there are no
+// more service account left with sunbscriber privileges, the topic will be
+// deleted. You can check this by calling ListSubscribers api.
+//
+// Possible Error Codes:
+//
+//   PERMISSION_DENIED: If the reseller account making the request and the
+//   reseller account being provided are different, or if the impersonated user
+//   is not a super admin.
+//
+//   INVALID_ARGUMENT: Missing or invalid required parameters in the
+//   request.
+//
+//   NOT_FOUND: If the topic resource doesn’t exist.
+//
+//   INTERNAL: Any non-user error related to a technical issue in the
+//   backend. In this case, contact Cloud Channel support.
+//
+//   UNKNOWN: Any non-user error related to a technical issue in
+//   the backend. In this case, contact Cloud Channel support.
+//
+// Return Value:
+// Topic name from which service email address has been unregistered if
+// successful, otherwise error is returned. If the service email was already
+// not associated with the topic, the success response will be returned.
+func (c *CloudChannelClient) UnregisterSubscriber(ctx context.Context, req *channelpb.UnregisterSubscriberRequest, opts ...gax.CallOption) (*channelpb.UnregisterSubscriberResponse, error) {
+	if _, ok := ctx.Deadline(); !ok && !c.disableDeadlines {
+		cctx, cancel := context.WithTimeout(ctx, 60000*time.Millisecond)
+		defer cancel()
+		ctx = cctx
+	}
+	md := metadata.Pairs("x-goog-request-params", fmt.Sprintf("%s=%v", "account", url.QueryEscape(req.GetAccount())))
+	ctx = insertMetadata(ctx, c.xGoogMetadata, md)
+	opts = append(c.CallOptions.UnregisterSubscriber[0:len(c.CallOptions.UnregisterSubscriber):len(c.CallOptions.UnregisterSubscriber)], opts...)
+	var resp *channelpb.UnregisterSubscriberResponse
+	err := gax.Invoke(ctx, func(ctx context.Context, settings gax.CallSettings) error {
+		var err error
+		resp, err = c.cloudChannelClient.UnregisterSubscriber(ctx, req, settings.GRPC...)
+		return err
+	}, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return resp, nil
+}
+
+// ListSubscribers lists service accounts with subscriber privileges on the Cloud Pub/Sub
+// topic created for this Channel Services account.
+//
+// Possible Error Codes:
+//
+//   PERMISSION_DENIED: If the reseller account making the request and the
+//   reseller account being provided are different, or if the account is not
+//   a super admin.
+//
+//   INVALID_ARGUMENT: Missing or invalid required parameters in the
+//   request.
+//
+//   NOT_FOUND: If the topic resource doesn’t exist.
+//
+//   INTERNAL: Any non-user error related to a technical issue in the
+//   backend. In this case, contact Cloud Channel support.
+//
+//   UNKNOWN: Any non-user error related to a technical issue in
+//   the backend. In this case, contact Cloud Channel support.
+//
+// Return Value:
+// List of service email addresses if successful, otherwise error is
+// returned.
+func (c *CloudChannelClient) ListSubscribers(ctx context.Context, req *channelpb.ListSubscribersRequest, opts ...gax.CallOption) *StringIterator {
+	md := metadata.Pairs("x-goog-request-params", fmt.Sprintf("%s=%v", "account", url.QueryEscape(req.GetAccount())))
+	ctx = insertMetadata(ctx, c.xGoogMetadata, md)
+	opts = append(c.CallOptions.ListSubscribers[0:len(c.CallOptions.ListSubscribers):len(c.CallOptions.ListSubscribers)], opts...)
+	it := &StringIterator{}
+	req = proto.Clone(req).(*channelpb.ListSubscribersRequest)
+	it.InternalFetch = func(pageSize int, pageToken string) ([]string, string, error) {
+		var resp *channelpb.ListSubscribersResponse
+		req.PageToken = pageToken
+		if pageSize > math.MaxInt32 {
+			req.PageSize = math.MaxInt32
+		} else {
+			req.PageSize = int32(pageSize)
+		}
+		err := gax.Invoke(ctx, func(ctx context.Context, settings gax.CallSettings) error {
+			var err error
+			resp, err = c.cloudChannelClient.ListSubscribers(ctx, req, settings.GRPC...)
+			return err
+		}, opts...)
+		if err != nil {
+			return nil, "", err
+		}
+
+		it.Response = resp
+		return resp.GetServiceAccounts(), resp.GetNextPageToken(), nil
 	}
 	fetch := func(pageSize int, pageToken string) (string, error) {
 		items, nextPageToken, err := it.InternalFetch(pageSize, pageToken)
@@ -2888,6 +3230,53 @@ func (it *SkuIterator) bufLen() int {
 }
 
 func (it *SkuIterator) takeBuf() interface{} {
+	b := it.items
+	it.items = nil
+	return b
+}
+
+// StringIterator manages a stream of string.
+type StringIterator struct {
+	items    []string
+	pageInfo *iterator.PageInfo
+	nextFunc func() error
+
+	// Response is the raw response for the current page.
+	// It must be cast to the RPC response type.
+	// Calling Next() or InternalFetch() updates this value.
+	Response interface{}
+
+	// InternalFetch is for use by the Google Cloud Libraries only.
+	// It is not part of the stable interface of this package.
+	//
+	// InternalFetch returns results from a single call to the underlying RPC.
+	// The number of results is no greater than pageSize.
+	// If there are no more results, nextPageToken is empty and err is nil.
+	InternalFetch func(pageSize int, pageToken string) (results []string, nextPageToken string, err error)
+}
+
+// PageInfo supports pagination. See the google.golang.org/api/iterator package for details.
+func (it *StringIterator) PageInfo() *iterator.PageInfo {
+	return it.pageInfo
+}
+
+// Next returns the next result. Its second return value is iterator.Done if there are no more
+// results. Once Next returns Done, all subsequent calls will return Done.
+func (it *StringIterator) Next() (string, error) {
+	var item string
+	if err := it.nextFunc(); err != nil {
+		return item, err
+	}
+	item = it.items[0]
+	it.items = it.items[1:]
+	return item, nil
+}
+
+func (it *StringIterator) bufLen() int {
+	return len(it.items)
+}
+
+func (it *StringIterator) takeBuf() interface{} {
 	b := it.items
 	it.items = nil
 	return b