+++ internal/godocfx/parse.go
@@ -14,6 +14,12 @@
 
 // +build go1.15
 
+// TODO:
+//   IDs for const/var groups have every name, not just the one to link to.
+//   Preserve IDs when sanitizing then use the right ID for linking.
+//   Link to different domains by pattern (e.g. for cloud.google.com/go).
+//   Make sure dot imports work (those identifiers aren't in the current package).
+
 package main
 
 import (
@@ -28,7 +34,9 @@ import (
 	"log"
 	"os"
 	"path/filepath"
+	"regexp"
 	"sort"
+	"strconv"
 	"strings"
 
 	"cloud.google.com/go/third_party/pkgsite"
@@ -148,7 +156,8 @@ func parse(glob string, workingDir string, optionalExtraFiles []string) (*result
 	// Once the files are grouped by package, process each package
 	// independently.
 	for _, pi := range pkgInfos {
-
+		link := newLinker(pi)
+		topLevelDecls := pkgsite.TopLevelDecls(pi.doc)
 		pkgItem := &item{
 			UID:      pi.doc.ImportPath,
 			Name:     pi.doc.ImportPath,
@@ -175,7 +184,7 @@ func parse(glob string, workingDir string, optionalExtraFiles []string) (*result
 				Type:    "const",
 				Summary: c.Doc,
 				Langs:   onlyGo,
-				Syntax:  syntax{Content: pkgsite.PrintType(pi.fset, c.Decl)},
+				Syntax:  syntax{Content: pkgsite.PrintType(pi.fset, c.Decl, link.toURL, topLevelDecls)},
 			})
 		}
 		for _, v := range pi.doc.Vars {
@@ -191,7 +200,7 @@ func parse(glob string, workingDir string, optionalExtraFiles []string) (*result
 				Type:    "variable",
 				Summary: v.Doc,
 				Langs:   onlyGo,
-				Syntax:  syntax{Content: pkgsite.PrintType(pi.fset, v.Decl)},
+				Syntax:  syntax{Content: pkgsite.PrintType(pi.fset, v.Decl, link.toURL, topLevelDecls)},
 			})
 		}
 		for _, t := range pi.doc.Types {
@@ -205,7 +214,7 @@ func parse(glob string, workingDir string, optionalExtraFiles []string) (*result
 				Type:     "type",
 				Summary:  t.Doc,
 				Langs:    onlyGo,
-				Syntax:   syntax{Content: pkgsite.PrintType(pi.fset, t.Decl)},
+				Syntax:   syntax{Content: pkgsite.PrintType(pi.fset, t.Decl, link.toURL, topLevelDecls)},
 				Examples: processExamples(t.Examples, pi.fset),
 			}
 			// Note: items are added as page.Children, rather than
@@ -224,7 +233,7 @@ func parse(glob string, workingDir string, optionalExtraFiles []string) (*result
 					Type:    "const",
 					Summary: c.Doc,
 					Langs:   onlyGo,
-					Syntax:  syntax{Content: pkgsite.PrintType(pi.fset, c.Decl)},
+					Syntax:  syntax{Content: pkgsite.PrintType(pi.fset, c.Decl, link.toURL, topLevelDecls)},
 				})
 			}
 			for _, v := range t.Vars {
@@ -240,7 +249,7 @@ func parse(glob string, workingDir string, optionalExtraFiles []string) (*result
 					Type:    "variable",
 					Summary: v.Doc,
 					Langs:   onlyGo,
-					Syntax:  syntax{Content: pkgsite.PrintType(pi.fset, v.Decl)},
+					Syntax:  syntax{Content: pkgsite.PrintType(pi.fset, v.Decl, link.toURL, topLevelDecls)},
 				})
 			}
 
@@ -255,7 +264,7 @@ func parse(glob string, workingDir string, optionalExtraFiles []string) (*result
 					Type:     "function",
 					Summary:  fn.Doc,
 					Langs:    onlyGo,
-					Syntax:   syntax{Content: pkgsite.Synopsis(pi.fset, fn.Decl)},
+					Syntax:   syntax{Content: pkgsite.Synopsis(pi.fset, fn.Decl, link.linkify)},
 					Examples: processExamples(fn.Examples, pi.fset),
 				})
 			}
@@ -270,7 +279,7 @@ func parse(glob string, workingDir string, optionalExtraFiles []string) (*result
 					Type:     "method",
 					Summary:  fn.Doc,
 					Langs:    onlyGo,
-					Syntax:   syntax{Content: pkgsite.Synopsis(pi.fset, fn.Decl)},
+					Syntax:   syntax{Content: pkgsite.Synopsis(pi.fset, fn.Decl, link.linkify)},
 					Examples: processExamples(fn.Examples, pi.fset),
 				})
 			}
@@ -286,7 +295,7 @@ func parse(glob string, workingDir string, optionalExtraFiles []string) (*result
 				Type:     "function",
 				Summary:  fn.Doc,
 				Langs:    onlyGo,
-				Syntax:   syntax{Content: pkgsite.Synopsis(pi.fset, fn.Decl)},
+				Syntax:   syntax{Content: pkgsite.Synopsis(pi.fset, fn.Decl, link.linkify)},
 				Examples: processExamples(fn.Examples, pi.fset),
 			})
 		}
@@ -298,6 +307,180 @@ func parse(glob string, workingDir string, optionalExtraFiles []string) (*result
 		module:     module,
 		extraFiles: extraFiles,
 	}, nil
+}
+
+type linker struct {
+	// imports is a map from local package name to import path.
+	// Behavior is undefined when a single import has different names in
+	// different files.
+	imports map[string]string
+
+	// idToAnchor is a map from package path to a map from ID to the anchor for
+	// that ID.
+	idToAnchor map[string]map[string]string
+
+	// sameDomainModules is a map from package path to module for every imported
+	// package that should cross link on the same domain.
+	sameDomainModules map[string]*packages.Module
+}
+
+func newLinker(pi pkgInfo) *linker {
+	sameDomainPrefixes := []string{"cloud.google.com/go"}
+
+	imports := map[string]string{}
+	sameDomainModules := map[string]*packages.Module{}
+	idToAnchor := map[string]map[string]string{}
+
+	for path, pkg := range pi.pkg.Imports {
+		name := pkg.Name
+		if rename := pi.importRenames[path]; rename != "" {
+			name = rename
+		}
+		imports[name] = path
+
+		// TODO: Consider documenting internal packages so we don't have to link
+		// out.
+		if pkg.Module != nil && hasPrefix(pkg.PkgPath, sameDomainPrefixes) && !strings.Contains(pkg.PkgPath, "internal") {
+			sameDomainModules[path] = pkg.Module
+
+			docPkg, _ := doc.NewFromFiles(pkg.Fset, pkg.Syntax, path)
+			idToAnchor[path] = buildIDToAnchor(docPkg)
+		}
+	}
+
+	idToAnchor[""] = buildIDToAnchor(pi.doc)
+
+	return &linker{imports: imports, idToAnchor: idToAnchor, sameDomainModules: sameDomainModules}
+}
+
+// nonWordRegex is based on
+// https://github.com/googleapis/doc-templates/blob/70eba5908e7b9aef5525d0f1f24194ae750f267e/third_party/docfx/templates/devsite/common.js#L27-L30.
+var nonWordRegex = regexp.MustCompile("\\W")
+
+func buildIDToAnchor(pkg *doc.Package) map[string]string {
+	idToAnchor := map[string]string{}
+	idToAnchor[pkg.ImportPath] = pkg.ImportPath
+
+	for _, c := range pkg.Consts {
+		commaID := strings.Join(c.Names, ",")
+		uid := pkg.ImportPath + "." + commaID
+		for _, name := range c.Names {
+			idToAnchor[name] = uid
+		}
+	}
+	for _, v := range pkg.Vars {
+		commaID := strings.Join(v.Names, ",")
+		uid := pkg.ImportPath + "." + commaID
+		for _, name := range v.Names {
+			idToAnchor[name] = uid
+		}
+	}
+	for _, f := range pkg.Funcs {
+		uid := pkg.ImportPath + "." + f.Name
+		idToAnchor[f.Name] = uid
+	}
+	for _, t := range pkg.Types {
+		uid := pkg.ImportPath + "." + t.Name
+		idToAnchor[t.Name] = uid
+		for _, c := range t.Consts {
+			commaID := strings.Join(c.Names, ",")
+			uid := pkg.ImportPath + "." + commaID
+			for _, name := range c.Names {
+				idToAnchor[name] = uid
+			}
+		}
+		for _, v := range t.Vars {
+			commaID := strings.Join(v.Names, ",")
+			uid := pkg.ImportPath + "." + commaID
+			for _, name := range v.Names {
+				idToAnchor[name] = uid
+			}
+		}
+		for _, f := range t.Funcs {
+			uid := pkg.ImportPath + "." + t.Name + "." + f.Name
+			idToAnchor[f.Name] = uid
+		}
+		for _, m := range t.Methods {
+			uid := pkg.ImportPath + "." + t.Name + "." + m.Name
+			idToAnchor[m.Name] = uid
+		}
+	}
+
+	for id, anchor := range idToAnchor {
+		idToAnchor[id] = nonWordRegex.ReplaceAllString(anchor, "_")
+	}
+
+	return idToAnchor
+}
+
+func (l *linker) linkify(s string) string {
+	prefix := ""
+	if strings.HasPrefix(s, "...") {
+		s = s[3:]
+		prefix = "..."
+	}
+	if s[0] == '*' {
+		s = s[1:]
+		prefix += "*"
+	}
+
+	if !strings.Contains(s, ".") {
+		// If s is not exported, it's probably a builtin.
+		if !token.IsExported(s) {
+			if doc.IsPredeclared(s) {
+				return href(l.toURL("builtin", s), s)
+			}
+			return fmt.Sprintf("%s%s", prefix, s)
+		}
+		return fmt.Sprintf("%s%s", prefix, href(l.toURL("", s), s))
+	}
+	// Otherwise, it's in another package.
+	split := strings.Split(s, ".")
+	if len(split) != 2 {
+		// Don't know how to link this.
+		return fmt.Sprintf("%s%s", prefix, s)
+	}
+
+	pkg := split[0]
+	pkgPath, ok := l.imports[pkg]
+	if !ok {
+		// Don't know how to link this.
+		return fmt.Sprintf("%s%s", prefix, s)
+	}
+	name := split[1]
+	return fmt.Sprintf("%s%s.%s", prefix, href(l.toURL(pkgPath, ""), pkg), href(l.toURL(pkgPath, name), name))
+}
+
+// TODO: link to the right baseURL, with the right module name and version
+// pattern.
+func (l *linker) toURL(pkg, name string) string {
+	if pkg == "" {
+		if anchor := l.idToAnchor[""][name]; anchor != "" {
+			name = anchor
+		}
+		return fmt.Sprintf("#%s", name)
+	}
+	if mod, ok := l.sameDomainModules[pkg]; ok {
+		pkgRemainder := ""
+		if pkg != mod.Path {
+			pkgRemainder = pkg[len(mod.Path)+1:] // +1 to skip slash.
+		}
+		// Note: we always link to latest. One day, we'll link to mod.Version.
+		baseURL := fmt.Sprintf("/go/docs/reference/%v/latest/%v", mod.Path, pkgRemainder)
+		if anchor := l.idToAnchor[pkg][name]; anchor != "" {
+			return fmt.Sprintf("%s#%s", baseURL, anchor)
+		}
+		return baseURL
+	}
+	baseURL := "https://pkg.go.dev"
+	if name == "" {
+		return fmt.Sprintf("%s/%s", baseURL, pkg)
+	}
+	return fmt.Sprintf("%s/%s#%s", baseURL, pkg, name)
+}
+
+func href(url, text string) string {
+	return fmt.Sprintf(`<a href="%s">%s</a>`, url, text)
 }
 
 // processExamples converts the examples to []example.
@@ -400,11 +583,13 @@ type pkgInfo struct {
 	pkg  *packages.Package
 	doc  *doc.Package
 	fset *token.FileSet
+	// importRenames is a map from package path to local name or "".
+	importRenames map[string]string
 }
 
 func loadPackages(glob, workingDir string) ([]pkgInfo, error) {
 	config := &packages.Config{
-		Mode:  packages.NeedName | packages.NeedSyntax | packages.NeedTypes | packages.NeedTypesInfo | packages.NeedModule,
+		Mode:  packages.NeedName | packages.NeedSyntax | packages.NeedTypes | packages.NeedTypesInfo | packages.NeedModule | packages.NeedImports | packages.NeedDeps,
 		Tests: true,
 		Dir:   workingDir,
 	}
@@ -440,7 +625,7 @@ func loadPackages(glob, workingDir string) ([]pkgInfo, error) {
 		}
 		if strings.Contains(id, "_test") {
 			id = id[0:strings.Index(id, "_test [")]
-		} else {
+		} else if pkg.Module != nil {
 			idToPkg[pkg.PkgPath] = pkg
 			pkgNames = append(pkgNames, pkg.PkgPath)
 			// The test package doesn't have Module set.
@@ -494,12 +679,38 @@ func loadPackages(glob, workingDir string) ([]pkgInfo, error) {
 			continue
 		}
 
+		imports := map[string]string{}
+		for _, f := range parsedFiles {
+			for _, i := range f.Imports {
+				name := ""
+				// i.Name is nil for imports that aren't renamed.
+				if i.Name != nil {
+					name = i.Name.Name
+				}
+				iPath, err := strconv.Unquote(i.Path.Value)
+				if err != nil {
+					return nil, fmt.Errorf("strconv.Unquote: %v", err)
+				}
+				imports[iPath] = name
+			}
+		}
+
 		result = append(result, pkgInfo{
-			pkg:  idToPkg[pkgPath],
-			doc:  docPkg,
-			fset: fset,
+			pkg:           idToPkg[pkgPath],
+			doc:           docPkg,
+			fset:          fset,
+			importRenames: imports,
 		})
 	}
 
 	return result, nil
+}
+
+func hasPrefix(s string, prefixes []string) bool {
+	for _, prefix := range prefixes {
+		if strings.HasPrefix(s, prefix) {
+			return true
+		}
+	}
+	return false
 }