+++ pubsublite/CHANGES.md
@@ -1,5 +1,35 @@
 # Changes
 
+## [0.7.0](https://www.github.com/googleapis/google-cloud-go/compare/v0.6.0...v0.7.0) (2021-02-18)
+
+The status of this library is now **BETA**.
+
+### Features
+
+* **pubsublite:** allow increasing the number of topic partitions ([#3647](https://www.github.com/googleapis/google-cloud-go/issues/3647)) ([1f85fdc](https://www.github.com/googleapis/google-cloud-go/commit/1f85fdca9f4317fab0f18b8bd9fcc8c65ab690e9))
+
+
+### Bug Fixes
+
+* **pubsublite:** change pubsub.Message.ID to an encoded publish.Metadata ([#3662](https://www.github.com/googleapis/google-cloud-go/issues/3662)) ([6b2807f](https://www.github.com/googleapis/google-cloud-go/commit/6b2807f1e13dc38eb79833f8d2766f27d4003434))
+* **pubsublite:** rebatch messages upon new publish stream ([#3694](https://www.github.com/googleapis/google-cloud-go/issues/3694)) ([0da3578](https://www.github.com/googleapis/google-cloud-go/commit/0da3578c8f007f71291cdc93d43f98acbe1dbb37))
+* **pubsublite:** rename publish.Metadata to pscompat.MessageMetadata ([#3672](https://www.github.com/googleapis/google-cloud-go/issues/3672)) ([6a8d4c5](https://www.github.com/googleapis/google-cloud-go/commit/6a8d4c515eb957d05e280e02e8cea9a89bdcbb1e))
+
+## [0.6.0](https://www.github.com/googleapis/google-cloud-go/compare/v0.5.0...v0.6.0) (2021-01-28)
+
+
+### ⚠ API Changes
+
+* **pubsublite:** add separate publisher and subscriber client constructors with settings ([#3528](https://www.github.com/googleapis/google-cloud-go/issues/3528)) ([98637e0](https://www.github.com/googleapis/google-cloud-go/commit/98637e089776292232bb7c039844680627ddade1))
+* **pubsublite:** rename package ps to pscompat ([#3569](https://www.github.com/googleapis/google-cloud-go/issues/3569)) ([9d8fd2b](https://www.github.com/googleapis/google-cloud-go/commit/9d8fd2b5e6999657bcf324878732da801b805591))
+* **pubsublite:** rename AdminClient TopicPartitions to TopicPartitionCount ([#3565](https://www.github.com/googleapis/google-cloud-go/issues/3565)) ([86a4de7](https://www.github.com/googleapis/google-cloud-go/commit/86a4de757bc2eed97577aba7fd51b5f5540e097e))
+* **pubsublite:** use strings for resource paths ([#3559](https://www.github.com/googleapis/google-cloud-go/issues/3559)) ([c18ed25](https://www.github.com/googleapis/google-cloud-go/commit/c18ed25900ba41e0b6b98a89cec8615df6a1146c))
+
+### Bug Fixes
+
+* **pubsublite:** close clients after publisher and subscriber have terminated ([#3512](https://www.github.com/googleapis/google-cloud-go/issues/3512)) ([72d2aff](https://www.github.com/googleapis/google-cloud-go/commit/72d2affb957cea7b6a223b108d0fe67c5635b25c))
+* **pubsublite:** ignore outstanding acks for unassigned partition subscribers ([#3597](https://www.github.com/googleapis/google-cloud-go/issues/3597)) ([eb91f1f](https://www.github.com/googleapis/google-cloud-go/commit/eb91f1f3c96f4c868e523f3c43f8c22b10ad4de4))
+
 ## [0.5.0](https://www.github.com/googleapis/google-cloud-go/compare/v0.4.0...v0.5.0) (2021-01-07)